requirements.txt
alembic==1.2.0
appdirs==1.4.3
aspy.yaml==1.3.0
attrs==19.2.0
black==19.3b0
cfgv==2.0.1
Click==7.0
docutils==0.15.2
dominate==2.4.0
elasticsearch==7.0.4
entrypoints==0.3
flake8==3.7.8
flake8-blind-except==0.1.1
flake8-builtins==1.4.1
flake8-docstrings==1.5.0
flake8-rst-docstrings==0.0.11
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
identify==1.4.7
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
nodeenv==1.3.3
pep8-naming==0.8.2
pre-commit==1.18.3
pycodestyle==2.5.0
pydocstyle==4.0.1
pyflakes==2.1.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
PyYAML==5.1.2
redis==3.3.8
restructuredtext-lint==1.3.0
rq==1.1.0
six==1.12.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.8
toml==0.10.0
urllib3==1.25.5
virtualenv==16.7.5
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
